+++ test/test_ephem_types.jl
@@ -1,12 +1,13 @@
 #=
 test_ephem_types.jl
 
-Unit tests for *Ephemerides types introduced in v0.2.0:
+Unit tests for parametric *Ephemerides{R} types introduced in v0.2.0:
 - Type hierarchy
-- Construction and field access
+- Construction via convenience constructors
+- Field access
 - Base.length
 - Inner constructor dimension validation
-- Upgrade constructors
+- Explicit parametric construction {Nothing} and {<:AbstractVector}
 =#
 
 @testset "Ephemerides types" begin
@@ -29,9 +30,9 @@ Unit tests for *Ephemerides types introduced in v0.2.0:
 
     @testset "Type hierarchy" begin
         ephem_s  = SingleAsteroidEphemerides(times, r_sun)
-        ephem_sd = SingleAsteroidEphemeridesWithDynamics(times, r_sun, R_body_to_inertial)
+        ephem_sd = SingleAsteroidEphemerides(times, r_sun, R_body_to_inertial)
         ephem_b  = BinaryAsteroidEphemerides(times, r_sun, r_secondary, R_primary_to_secondary)
-        ephem_bd = BinaryAsteroidEphemeridesWithDynamics(times, r_sun, r_secondary, R_primary_to_secondary, R_primary_to_inertial)
+        ephem_bd = BinaryAsteroidEphemerides(times, r_sun, r_secondary, R_primary_to_secondary, R_primary_to_inertial)
 
         @test ephem_s  isa AbstractSingleAsteroidEphemerides
         @test ephem_sd isa AbstractSingleAsteroidEphemerides
@@ -42,56 +43,58 @@ Unit tests for *Ephemerides types introduced in v0.2.0:
         @test ephem_sd isa AbstractAsteroidEphemerides
         @test ephem_b  isa AbstractAsteroidEphemerides
         @test ephem_bd isa AbstractAsteroidEphemerides
+
+        # Parametric type parameters
+        @test ephem_s  isa SingleAsteroidEphemerides{Nothing}
+        @test ephem_sd isa SingleAsteroidEphemerides{Vector{SMatrix{3,3,Float64,9}}}
+        @test ephem_b  isa BinaryAsteroidEphemerides{Nothing}
+        @test ephem_bd isa BinaryAsteroidEphemerides{Vector{SMatrix{3,3,Float64,9}}}
     end
 
-    @testset "SingleAsteroidEphemerides" begin
+    @testset "SingleAsteroidEphemerides{Nothing}" begin
         ephem = SingleAsteroidEphemerides(times, r_sun)
 
-        @test ephem.times === times
-        @test ephem.r_sun === r_sun
+        @test ephem.times              === times
+        @test ephem.r_sun              === r_sun
+        @test ephem.R_body_to_inertial === nothing
         @test length(ephem) == n
 
         # DimensionMismatch
-        @test_throws DimensionMismatch SingleAsteroidEphemerides(times[1:end-1], r_sun)
+        @test_throws DimensionMismatch SingleAsteroidEphemerides{Nothing}(times[1:end-1], r_sun, nothing)
     end
 
-    @testset "SingleAsteroidEphemeridesWithDynamics" begin
-        ephem = SingleAsteroidEphemeridesWithDynamics(times, r_sun, R_body_to_inertial)
+    @testset "SingleAsteroidEphemerides{Vector{SMatrix}}" begin
+        ephem = SingleAsteroidEphemerides(times, r_sun, R_body_to_inertial)
 
         @test ephem.times              === times
         @test ephem.r_sun              === r_sun
         @test ephem.R_body_to_inertial === R_body_to_inertial
         @test length(ephem) == n
 
-        @test_throws DimensionMismatch SingleAsteroidEphemeridesWithDynamics(times, r_sun[1:end-1], R_body_to_inertial)
-        @test_throws DimensionMismatch SingleAsteroidEphemeridesWithDynamics(times, r_sun, R_body_to_inertial[1:end-1])
-
-        # Upgrade constructor
-        ephem_base = SingleAsteroidEphemerides(times, r_sun)
-        ephem_up   = SingleAsteroidEphemeridesWithDynamics(ephem_base, R_body_to_inertial)
-
-        @test ephem_up isa SingleAsteroidEphemeridesWithDynamics
-        @test ephem_up.times === ephem_base.times
-        @test ephem_up.r_sun === ephem_base.r_sun
-        @test ephem_up.R_body_to_inertial === R_body_to_inertial
+        # DimensionMismatch
+        @test_throws DimensionMismatch SingleAsteroidEphemerides(times, r_sun[1:end-1], R_body_to_inertial)
+        @test_throws DimensionMismatch SingleAsteroidEphemerides(times, r_sun, R_body_to_inertial[1:end-1])
     end
 
-    @testset "BinaryAsteroidEphemerides" begin
+    @testset "BinaryAsteroidEphemerides{Nothing}" begin
         ephem = BinaryAsteroidEphemerides(times, r_sun, r_secondary, R_primary_to_secondary)
 
         @test ephem.times                  === times
         @test ephem.r_sun                  === r_sun
         @test ephem.r_secondary            === r_secondary
         @test ephem.R_primary_to_secondary === R_primary_to_secondary
+        @test ephem.R_primary_to_inertial  === nothing
         @test length(ephem) == n
 
+        # DimensionMismatch
+        @test_throws DimensionMismatch BinaryAsteroidEphemerides{Nothing}(times[1:end-1], r_sun, r_secondary, R_primary_to_secondary, nothing)
         @test_throws DimensionMismatch BinaryAsteroidEphemerides(times, r_sun[1:end-1], r_secondary, R_primary_to_secondary)
         @test_throws DimensionMismatch BinaryAsteroidEphemerides(times, r_sun, r_secondary[1:end-1], R_primary_to_secondary)
         @test_throws DimensionMismatch BinaryAsteroidEphemerides(times, r_sun, r_secondary, R_primary_to_secondary[1:end-1])
     end
 
-    @testset "BinaryAsteroidEphemeridesWithDynamics" begin
-        ephem = BinaryAsteroidEphemeridesWithDynamics(times, r_sun, r_secondary, R_primary_to_secondary, R_primary_to_inertial)
+    @testset "BinaryAsteroidEphemerides{Vector{SMatrix}}" begin
+        ephem = BinaryAsteroidEphemerides(times, r_sun, r_secondary, R_primary_to_secondary, R_primary_to_inertial)
 
         @test ephem.times                  === times
         @test ephem.r_sun                  === r_sun
@@ -100,20 +103,10 @@ Unit tests for *Ephemerides types introduced in v0.2.0:
         @test ephem.R_primary_to_inertial  === R_primary_to_inertial
         @test length(ephem) == n
 
-        @test_throws DimensionMismatch BinaryAsteroidEphemeridesWithDynamics(times, r_sun[1:end-1], r_secondary, R_primary_to_secondary, R_primary_to_inertial)
-        @test_throws DimensionMismatch BinaryAsteroidEphemeridesWithDynamics(times, r_sun, r_secondary[1:end-1], R_primary_to_secondary, R_primary_to_inertial)
-        @test_throws DimensionMismatch BinaryAsteroidEphemeridesWithDynamics(times, r_sun, r_secondary, R_primary_to_secondary[1:end-1], R_primary_to_inertial)
-        @test_throws DimensionMismatch BinaryAsteroidEphemeridesWithDynamics(times, r_sun, r_secondary, R_primary_to_secondary, R_primary_to_inertial[1:end-1])
-
-        # Upgrade constructor
-        ephem_base = BinaryAsteroidEphemerides(times, r_sun, r_secondary, R_primary_to_secondary)
-        ephem_up   = BinaryAsteroidEphemeridesWithDynamics(ephem_base, R_primary_to_inertial)
-
-        @test ephem_up isa BinaryAsteroidEphemeridesWithDynamics
-        @test ephem_up.times                  === ephem_base.times
-        @test ephem_up.r_sun                  === ephem_base.r_sun
-        @test ephem_up.r_secondary            === ephem_base.r_secondary
-        @test ephem_up.R_primary_to_secondary === ephem_base.R_primary_to_secondary
-        @test ephem_up.R_primary_to_inertial  === R_primary_to_inertial
+        # DimensionMismatch
+        @test_throws DimensionMismatch BinaryAsteroidEphemerides(times, r_sun[1:end-1], r_secondary, R_primary_to_secondary, R_primary_to_inertial)
+        @test_throws DimensionMismatch BinaryAsteroidEphemerides(times, r_sun, r_secondary[1:end-1], R_primary_to_secondary, R_primary_to_inertial)
+        @test_throws DimensionMismatch BinaryAsteroidEphemerides(times, r_sun, r_secondary, R_primary_to_secondary[1:end-1], R_primary_to_inertial)
+        @test_throws DimensionMismatch BinaryAsteroidEphemerides(times, r_sun, r_secondary, R_primary_to_secondary, R_primary_to_inertial[1:end-1])
     end
 end